+++ backend/inpainting-service/pyproject.toml
@@ -27,12 +27,8 @@ dev = ["pytest>=7.0.0", "black>=23.0.0"]
 aibox-image-lib = { path = "../aibox-image-lib", editable = true }
 
 [project.scripts]
-service = "service:main"
-
-[tool.black]
-line-length = 88
-target-version = ['py39']
+service = "inpainting_service.main:main"
 
 [build-system]
-requires = ["setuptools>=61.0", "wheel"]
-build-backend = "setuptools.build_meta"
+requires = ["uv_build>=0.8.15,<0.9.0"]
+build-backend = "uv_build"
