+++ backend/selection-service/pyproject.toml
@@ -5,18 +5,6 @@ description = "Image area selection service"
 requires-python = ">=3.9,<3.12"
 dependencies = [
     "aibox-image-lib",
-    # # Core dependencies
-    # "numpy>=1.21.0",
-    # "pillow>=10.0.0",
-    # IPC communication
-    "pyzmq>=25.0.0",
-    "msgpack>=1.0.0",
-    # # LLM and Vision models (needs latest transformers for SigLIP2)
-    # "torch>=2.4.0",
-    # "transformers>=4.53.3", # Latest for SigLIP2 support
-    # "accelerate>=1.10.0",
-    # # Text embedding models
-    # "sentence-transformers>=2.7.0", # For Qwen3-Embedding-0.6B
     "numpy>=2.0.2",
     "zmq>=0.0.0",
     "torch>=2.8.0",
@@ -34,10 +22,6 @@ dev = ["pytest>=7.0.0", "black>=23.0.0"]
 
 [project.scripts]
 service = "selection_service.main:main"
-
-[tool.black]
-line-length = 88
-target-version = ['py39']
 
 [build-system]
 requires = ["uv_build>=0.8.15,<0.9.0"]
